requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
fastapi==0.115.13
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
msal==1.32.3
msal-extensions==1.3.1
openai==1.91.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
